requirements.txt
# Core dependencies
Django>=4.2.0,<5.3.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.3.0,<5.0.0
django-filter>=23.5,<24.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0  # For PostgreSQL

# PDF generation
reportlab>=4.0.7,<5.0.0

# Security
cryptography>=41.0.0,<42.0.0

# Development
django-debug-toolbar>=4.2.0,<5.0.0
black>=23.11.0,<24.0.0
flake8>=6.1.0,<7.0.0

# Production
gunicorn>=21.2.0,<22.0.0

# Forms
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=2022.12 # Or your preferred crispy forms template pack

# External APIs
requests>=2.31.0,<3.0.0
